requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
black==22.12.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Migrate==3.1.0
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
PyJWT==2.1.0
pyrsistent==0.18.0
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.21.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1